server/requirements.txt
alembic==1.13.1
aniso8601==9.0.1
bcrypt==4.1.2
bidict==0.23.1
blinker==1.7.0
blueprint==3.4.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cloudinary==1.41.0
distlib==0.3.8
Faker==24.9.0
filelock==3.13.1
Flask==3.0.3
Flask-Bcrypt==1.0.1
flask-blueprint==1.3.0
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.0.7
Flask-RESTful==0.3.10
Flask-SocketIO==5.3.7
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
gunicorn==21.2.0
h11==0.14.0
honcho==1.1.0
idna==3.7
importlib_metadata==7.1.0
importlib_resources==6.4.0
itsdangerous==2.1.2
Jinja2==3.1.3
jsonify==0.5
make_response==1
Mako==1.3.3
MarkupSafe==2.1.5
packaging==24.0
pipenv==2023.12.1
platformdirs==4.2.0
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.9.1
python-socketio==5.11.4
pytz==2024.1
requests==2.32.2
simple-websocket==1.0.0
six==1.16.0
SQLAlchemy==2.0.29
SQLAlchemy-serializer==1.4.12
typing_extensions==4.11.0
urllib3==2.2.1
virtualenv==20.25.1
Werkzeug==3.0.2
wsproto==1.2.0
zipp==3.18.1
